requirements.txt
# Core dependencies
anthropic>=0.7.0
pydantic>=2.0.0
PyYAML>=6.0
python-dotenv>=1.0.0

# Data sources
yfinance>=0.2.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
aiohttp>=3.9.0

# Technical analysis (pure Python, no C deps)
pandas-ta>=0.3.14b

# OpenTelemetry
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0

# Async support
aiofiles>=23.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
